v1: fix `NaN` pollution in exponential backoff logic
In `delayForAttempt()`, when `durf` come from `math.Pow()` is `Inf`,
`jitterDelta` also becomes `Inf` and `jitteredMin` becomes `NaN` because
the result of `Inf - Inf` is `NaN`.
This `NaN` pollutes the above `durf` so `time.Duration(NaN)` will be a
result of this function.
This might not be expected (see also: https://play.golang.org/p/s3nG-s4Z9En)
and the backoff mechanism won't work correctly.