Fix the jitter calculation in the exponential policy
Until this commit, if the interval value is `1` (1 ns) and jitter factor is `0.5` (i.e. 50%),
a values range becomes `[0.5, 2.5)`. This range is not within 50% variation.
This issue is relieved if the interval value is big, but on the other hand, the issue
is remarkable if the value is small.
This commit removes unnecessary `+1` from random value derivation.